Add memory for output and register setpoint
Class definition MAX5134 now includes setpoint and register, which serve as a local storage for what the output/register of an instance is set to. Since most applications will use the setpoint and then measure something else, it is helpful to know what the DAC is set to without finding the code that set it. Though this does not detect any hardware errors that may occur.